+++ src/App.js
@@ -8,140 +8,9 @@ import './App.color.css';
 import unitToIdol  from "./unitToIdol.json";
 import idolToUnit  from "./idolToUnit.json";
 import unitName    from "./unitName.json"
-import idolData from "./idolData.json";
 
-class IdolButton extends React.Component {
-  render() {
-    var style = {
-      borderColor: idolData[this.props.idol].color
-    }
-    if(this.props.selectedIdols.indexOf(this.props.idol) !== -1){
-      style["backgroundColor"] = idolData[this.props.idol].color
-    }
-
-    return (
-      <div
-        className={`idolBox${this.props.classNameSuffix}`}
-        id={this.props.idol}
-        style={style} onClick={()=>{
-          this.props.onClickHandler(this.props.idol);
-        }}
-      >
-        <div className="boxContainer">
-          <div className={`idolImage${this.props.classNameSuffix} idolIcon-${this.props.idol}`}></div>
-          <div className={`idolNameBox${this.props.classNameSuffix}`}>
-            {idolData[this.props.idol].screenName}
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class IdolsSelect extends React.Component {
-  constructor(props){
-    super(props)
-    this.state={
-      filterType: ["princess", "fairy", "angel"],
-    }
-  }
-
-  setFilter(type){
-    if(this.state.filterType.indexOf(type) !== -1){
-      const filterType = this.state.filterType.filter(v=>v!==type)
-      this.setState({filterType: filterType})
-    }else{
-      const filterType = this.state.filterType
-      filterType.push(type)
-      this.setState({filterType: filterType})
-    }
-  }
-  
-  filterButtonBuilder(type){
-    return(
-      <label>
-        <div
-          className={`filterCheckboxType filterCheckbox-${type} ${(this.state.filterType.indexOf(type)!==-1)?"checked":""}`}
-        >
-          <div className="typeNameBox">
-            <input
-              type="checkbox"
-              checked={(this.state.filterType.indexOf(type)!==-1)?"checked":""}
-              onChange={()=>this.setFilter(type)}
-            />
-            {type}
-          </div>
-        </div>
-      </label>
-    )
-  }
-
-  render() {
-    const buttons = this.props.idols.map((idol, index) => {
-      if(this.state.filterType.indexOf(idolData[idol].type) === -1){
-        if(this.state.filterType.indexOf("selecting") === -1 ||
-           this.props.selectedIdols.indexOf(idol) === -1){
-          return null
-        }
-      }
-      return(
-        <IdolButton
-          key={index}
-          selectedIdols={this.props.selectedIdols}
-          idol={idol}
-          classNameSuffix={this.props.classNameSuffix}
-          onClickHandler={(idol)=>this.props.onClickHandler(idol)}
-        />
-      );
-    });
-    return (
-      <div className="idolView">
-        <form>
-          {this.filterButtonBuilder("princess")}
-          {this.filterButtonBuilder("fairy")}
-          {this.filterButtonBuilder("angel")}
-          {this.filterButtonBuilder("selecting")}
-        </form>
-        {buttons}
-      </div>
-    );
-  }
-}
-
-class UnitItem extends React.Component {
-  render(){
-    const unitMember = unitToIdol[unitName[this.props.unit]]
-    unitMember.sort((a,b)=>{
-      return (this.props.selectedIdols.indexOf(b)!==-1)?1:-1
-    })
-    const memberButtons = unitMember.map((member, i) => {
-      return(
-        <IdolButton
-          key={i}
-          selectedIdols={this.props.selectedIdols}
-          idol={member}
-          classNameSuffix={this.props.classNameSuffix}
-          onClickHandler={(idol)=>this.props.toggleClickHandler(idol)}
-        />
-      )
-    });
-    const selectedMember = unitMember.filter(idol=>this.props.selectedIdols.indexOf(idol)!==-1)
-    const completeClassName = (selectedMember.length === unitMember.length)?"unitComplete":""
-
-    return(
-      <div key={this.props.unit} className={`unitTable${this.props.classNameSuffix} ${completeClassName}`}>
-        <div className={`unitNameView${this.props.classNameSuffix}`}>
-          <div className={`unitNameBox${this.props.classNameSuffix}`}>
-            {unitName[this.props.unit]}
-          </div>
-        </div>
-        <div className="unitIdolView">
-          {memberButtons}
-        </div>
-      </div>
-    )
-  }
-}
+import IdolsSelect from './components/IdolsSelect';
+import UnitItem from './components/UnitItem';
 
 class App extends React.Component {
   constructor(props){
